order/src/events/listeners/ticket-updated-listener.ts
import {
  Listener,
  NotFoundError,
  Subjects,
  TicketUpdatedEvent,
} from "@db97tickets/common";
import { queueGroupName } from "./queue-group-name";
import { Message } from "node-nats-streaming";
import { Ticket } from "../../models/ticket.model";

export class TicketUpdatedListener extends Listener<TicketUpdatedEvent> {
  readonly subject = Subjects.TicketUpdated;
  queueGroupName = queueGroupName;

  async onMessage(
    data: TicketUpdatedEvent["data"],
    msg: Message
  ): Promise<void> {
    const ticket = await Ticket.findByIdAndPreviousVersion({
      id: data.id,
      version: data.version,
    });

    if (!ticket) {
      throw new NotFoundError("Ticket not found");
    }

    const { title, price } = data;
    ticket.set({ title, price });
    await ticket.save();

    msg.ack();
  }
}
